# docs/swagger/Swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: Restful Monitor
  contact: {}
host: example.com
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /registration:
    post:
      summary: registration
      tags:
        - users
      operationId: registration
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: accept-language
          in: header
          required: true
          type: string
          description: ""
        - name: Body
          in: body
          required: true
          description: ""
          schema:
            $ref: "#/definitions/registrationrequest"
      responses:
        "200":
          description: ""
          headers: {}
...
definitions:
...
  registrationrequest:
    title: registrationrequest
    example:
      email: omar.alsahragty@gmail.com
    type: object
    properties:
      email:
        type: string
    required:
      - email
...